emails/requirements.txt
python-dotenv
sendgrid
flask